+++ docs/CHANGELOG.md
@@ -8,7 +8,7 @@
 
 ### Added
 
-* new `HttpAgent` option: `delayStrategy` - allows you to set a custom delay strategy for retries. The default is a newly exported `exponentialBackoff`, but you can pass your own function to customize the delay between retries.
+* new `HttpAgent` option: `backoffStrategy` - allows you to set a custom delay strategy for retries. The default is a newly exported `exponentialBackoff`, but you can pass your own function to customize the delay between retries.
 
 ## [1.2.0] - 2024-03-25